Fix condition in pot-file-update workflow
The current condition is wrong. The new pot file is updated only in case
that the only `POT-Creation-Date` has changed, however, the condition
should work opposite. The fix is simple switch the condition by removing
`!` negation.

This issue unfortunately caused us to have old translations for last 4
months for most of the branches.